Fix Compass Panel, Add in Signals to scan system, #2597 mat updates scan
Compass panel formatting reworked to now work in high DPI
Added in signals to star scan class, scan display now shows them
Add back missing tooltips to roll up panels
_signals added to Action Scan command
Fix 2597 materials not updating scan display
Travelhistory double click - opens window now more, based on number of lines vs data grid view height
Reformatted some of journal scan description output to remove blank lines and move terraforming higher.
Star scan in its own folder, broken out more (its v.complicated) and added in SAASignalsFound.
