+++ EliteDangerous/EliteDangerous.csproj
@@ -99,6 +99,8 @@
     <Compile Include="Journal\JournalEssentialEvents.cs" />
     <Compile Include="Journal\JournalEntryDB.cs" />
     <Compile Include="Journal\JournalEnums.cs" />
+    <Compile Include="ScarScan\StarScanHelpers.cs" />
+    <Compile Include="ScarScan\StarScanDesignationMaps.cs" />
     <Compile Include="ScarScan\StarScanJournalScans.cs" />
     <Compile Include="ScarScan\StarScanBodyID.cs" />
     <Compile Include="ScarScan\StarScanSAASignalsFound.cs" />
